Create String-Corrector-Algorithm.c
Implementation of String-Corrector Algorithm. It involves getting symbol from the user. If the symbol is an alphabet and if it is a lower case then it is converted to upper case and vice verser. If the symbol is a digit or a punctuation it is discarded. If it is a white space character a space character is inserted. Until the character is different than the newline character and queue is not empty, the character is removed from the queue. The output will now be displayed on the screen.